// libs/wui/src/enum/WuiColorAlias.enum.ts
// Do not edit directly, this file was auto-generated.
// Generated from tokens/semantic/color.json

export enum WuiColorAlias {
  DANGER_0 = "danger-0",
  DANGER_100 = "danger-100",
  DANGER_200 = "danger-200",
  DANGER_300 = "danger-300",
  DANGER_400 = "danger-400",
  DANGER_500 = "danger-500",
  DANGER_600 = "danger-600",
  DANGER_700 = "danger-700",
  DANGER_800 = "danger-800",
  DANGER_900 = "danger-900",
  INFO_0 = "info-0",
  INFO_100 = "info-100",
  INFO_200 = "info-200",
  INFO_300 = "info-300",
  INFO_400 = "info-400",
  INFO_500 = "info-500",
  INFO_600 = "info-600",
  INFO_700 = "info-700",
  INFO_800 = "info-800",
  INFO_900 = "info-900",
  NEUTRAL_0 = "neutral-0",
  NEUTRAL_100 = "neutral-100",
  NEUTRAL_200 = "neutral-200",
  NEUTRAL_300 = "neutral-300",
  NEUTRAL_400 = "neutral-400",
  NEUTRAL_500 = "neutral-500",
  NEUTRAL_600 = "neutral-600",
  NEUTRAL_700 = "neutral-700",
  NEUTRAL_800 = "neutral-800",
  NEUTRAL_900 = "neutral-900",
  SUCCESS_0 = "success-0",
  SUCCESS_100 = "success-100",
  SUCCESS_200 = "success-200",
  SUCCESS_300 = "success-300",
  SUCCESS_400 = "success-400",
  SUCCESS_500 = "success-500",
  SUCCESS_600 = "success-600",
  SUCCESS_700 = "success-700",
  SUCCESS_800 = "success-800",
  SUCCESS_900 = "success-900",
  WARNING_0 = "warning-0",
  WARNING_100 = "warning-100",
  WARNING_200 = "warning-200",
  WARNING_300 = "warning-300",
  WARNING_400 = "warning-400",
  WARNING_500 = "warning-500",
  WARNING_600 = "warning-600",
  WARNING_700 = "warning-700",
  WARNING_800 = "warning-800",
  WARNING_900 = "warning-900"
}